Exclude files matching pattern "._text"
Hi Frenck, 

One change I routinely make after updating VSCode is to exclude the ._text.yaml files from the side bar.  The reason is these files are not valuable for updating Home Assistants configuration.  I am not sure, but it seems that this might be unique to OSX.  Link that I followed: https://stackoverflow.com/questions/30140112/how-do-i-hide-certain-files-from-the-sidebar-in-visual-studio-code

If you think this would help others, please consider merging.